Output all broadcasts from the source chosen for broadcasts
In the case that there are multiple items from the source
whose broadcasts are chosen as precedent, all broadcasts
from that source should be output, not just those from the
most precedent item.

It is assumed that within a data source, broadcasts aren't
duplicated, hence this is a valid approach. Otherwise, in the
case of multiple items from the same source being in an
equivalent set, some broadcasts will, erroneously, be omitted.
One such case is Press Association which may have generic
episodes for a show, alongside full hierarchical data.