Change ES deduping to take duplicate with highest precedence publisher
- Currently we grab the ID we see out of every equivalent set. This
creates the problem that after merging on output we might show
different values on the API than the ones used for ordering by ES
thus the ordering will appear wrong. This is intended to fix that
- Change IndexQueryResult to only store IDs and create separate
DelegateIndexQueryResult to store additional information required by
the PseudoEquivalentContentIndex (namely canonicalID and publisher) to
avoid leaking that information to other parts of the code that don't
need them and making those interfaces more confusing. The interface
DelegateContentIndex has been created to handle communication between
the PseudoEquivalentContentIndex and EsUnequivalentContentIndex
- Have PseudoEquivalentContentIndex return the normal ID of the
content chosen after deduping. This is better than returning the
canonicalId since--deduping aside--the ES store does not store
equivalent sets and therefore should not be returning a canonical ID.
It also eliminates several issues we've had caused by inconsistencies
of canonical ID between the ES store and the equivalent content store