Do not resolve equivalence graphs from stale index entries
- Due to past bugs we could have stale index entries that point to an
equivalence graph that does not actually contain the resource ID in
question. This will prevent resolution of the graph from the stale
index and will instead return absent, i.e. that a graph has not been
found for the requested resource ID
- Also this removes the code in AbstractEquivalenceGraphStore that
was attempting to handle this issue by deleting the "corrupt" graph
because the graph that is resolved from the stale index is not
necessarily a bad one. At that point all we know is that it should
not have been resolved for that particular requested resource ID