Implement integration tests for CQL content store
Also stop dropping the keyspace for C* integration tests.

- Switch dropping the keyspace to truncation for clearing out the DB
- Switch all ITs to doing CREATE IF NOT EXISTS operations rather than a full create

Schema operations are very slow in C*, CREATE IF NOT EXISTS seems to be much faster.
Since we were dropping the keyspace in fixture tear-down and recreating in setup, this
was making the ITs run very slow. With the introduction of the new CQL content store,
we add a lot of UDTs, which all need to be created, and this then effectively doubled
the IT running time.