Adding reporting for content migration so that we know more about content migration.
Updating the Jackson version to 2.8

Adding annotations to Id class so that we can get id as long and Long. Updated common dependency to the new version 2 as Timestamp millis method now also have Jackson annotations that we need in the deserialization logic.

Fixing failing tests as the deserialization logic for Joda intervals has changed and when deserializing interval, it's automatically converted to UTC.