Change channel_group endpoint to not show channel groups of channels
Currently this call is very expensive when used with the 'channel_groups_summary' annotation.

It returns all channels belonging to each returned channel group and also for each channel, every channel group that it belongs to.
This change removes the writing of the last part when the /channel_group/ endpoint is called.

The previous behaviour has been retained for the /channel/ endpoint.