Move resolution of region annotation for channel groups to channel group executor
In order to move the resolution, a composite object had to be introduced to hold
the extra data resolved from the executor to pass through to the writers.

A new writer has been introduced to write the field for ID regardless of its
nested context.

Many small changes have been made to method signatures as a generic type was
maintained throughout the entire request -> executor -> write pipeline. This
shouldn't actually affect the pipeline as there is no logic that changes its
function based on the type.

This is the basis for the rest of the annotations to follow suite in the
same format laid out in this commit.