Remove audit timestamps from indirect CQL row clustering keys
Prior to this we had a number of things as, e.g., CQL sets. Those things had the usual Atlas timestamps in them, e.g., lastUpdated,
which then resulted in the same item getting added multiple times, because the timestamp had changed. Most of these cases have now been
moved to CQL maps with the unique parts extracted to the key.