Hardcoded list of ID's to filter by before resolution
The desired effect was there which was removing channel groups
that were not needed however there was no performance gain as
all the channel groups still had to be resolved so we could
check their aliases. A list from the db has been hardcoded so now
we can filter before the ID's are passed to the resolver.