Whitelist bt channels by instance of ChannelGroupMembership (#435)
When creating a ChannelGroupMembership for a channel, if it has a number, it will be ChannelNumbering subclass of ChannelGroupMembership. As special channel groups do not have numbers, they are all ChannelGroupMemberships meaning we can filter out the instances of ChannelNumbering to remove the channel groups we do not want for that annotation.

Both annotations have now been deprecated as channel_groups_summary causes performance issues and generic_channel_groups_summary cannot be implemented in the desired way. These annotations should only be used by existing users of them.