Remove unnecessary test override and rename ItemRef UDT to PartialItemRef
`ItemRef` only contained the non-PK fields and was supposed to be used only in conjunction with `Ref` anyway. Therefore to
avoid having nameclashes if we ever want a full `ItemRef` object in the future, this one was renamed to be clearer.

As for the test, in classic late-Friday-night fashion I implemented it even though Thomas had already done it in the parent
IT class. Brain farts, plain and simple.