Make CQL migrate endpoints use full store instead of non-validating one
The non-validating one wasn't writing the denormalisations we have, like itemSummaries and itemRefs,
and it was impossible to fix it whilst maintaining the prime principle of its existence -- not doing
read-modify-writes, so therefore we're now doing everything via the regular store. The "non-validating"
one was taken out back and put down in a perfectly humane way.