Changed default behaviour of sendKafkaMessages in ContentDebugController
This is a temporary change due to the number of kafka messages
being produced being higher than expected. Since the endpoint
is being hit automatically within a lot of BARB processing
jobs it is causing kakfa queues to grow with what appears
to be a lot of duplicate messages. This suggests there may
be an issue with migration logic determining whether content
has actually changed.