Fix content hashing logic used in CqlContentStore
The issue comes in two parts:

1. The hasher class does not ignore certain date
fields which are always different between Mongo
and Cassandra. For example the last_updated field
is changed before writing to Cassandra and so
will generally always be different from the field
in Mongo.
For such fields that do not seem relevant to the
hashing logic and likely to be different, an
ExcludeFromHash annotation has been added.

2. The hashing check is performed too early in
the CqlContentStore.
The check has been moved to just before the content
is serialized, since it is modified in several places
before then.