Potentially fixes ENG-726
This causes race conditions when we try to read the graph before it is there
while data are there. As we are writting in a batch, switching the order
should ensure the graph will be there if the data are there. However, I
believe replication lag would be a bigger driver of this and switching the
order of the executeAsync will probably not make much difference, but worth a try.