Remove decoding step in resolving content from Sherlock. (#654)
Sherlock now uses decoded ids for its documents, so this prevents an error that was occuring where the decoding step in the ContentResolvingSearcher was finding illegal characters in the id string (because they were already decoded). Also moved the idCodec back to QueryWebModule where it was before adding Sherlock to deer, as there is no longer any use for it in the QueryModule.