Improve channel resolving in broadcasts annotations (#659)
Instead of resolving the channel of each broadcast one at a time the annotations now form a set of channelIds to resolve, stores them in a map and then fetches from the map as needed. For content with a large number of broadcasts this should remove a rather large bottleneck.

The aggregated broadcasts annotation hasn't been changed although it probably needs to be, because the change required is more complicated due to the structure of the logic.