allows channels from all sources to be merged
This is done to support to the BT MPX channels that equiv to the
new YV channels (instead of the existing PA channels that were
on the MB source). We enabled it for channels as the only channels
we know of that would be merged are barb and PA and BT and PA, so
really no harm in merging those, especially since the only attributes
that are being merged are aliases and dates.

In general it feels the merging was done to support specifically  the
BT/PA work, and it isn't a fully fledged/though after solution like
content merging, so more thought needs to go into this if we need to
support proper merging.