Image merging rework (ENG-1023) (#678)
The images set now includes all images from all equivs since it was a promised feature and to make it more consistent with merging of other collection fields.

If the content's image field is not set or is an unavailable image, an available image can now be chosen for it from the images set. In particular this allows another image from the same source to take the place of the expired image, instead of a lower precedent source image being chosen.

Unavailable images are now excluded from the images set as well unless a new all_images annotation is used.