Make force migration endpoint also force equiv updates
In addition to forcing content updates.

It has been observed that outgoing equiv links do not always end up as incoming equiv links on the corresponding piece of content when equivs are migrated. This endpoint allows us to fix such cases as we discover them until the root cause is solved.